+++ lib/MIP/Cli/Mip/Analyse/Rd_dna_panel.pm
@@ -17,7 +17,7 @@ use Moose::Util::TypeConstraints;
 ## MIPs lib
 use MIP::Main::Analyse qw{ mip_analyse };
 
-our $VERSION = 1.05;
+our $VERSION = 1.06;
 
 extends(qw{ MIP::Cli::Mip::Analyse });
 
@@ -1330,7 +1330,7 @@ q{Default: hgvs, symbol, numbers, sift, polyphen, humdiv, domains, protein, ccds
     option(
         q{qccollect_eval_metric_file} => (
             cmd_aliases   => [qw{ qccemf }],
-            cmd_tags      => [q{Default: qc_eval_metric_-v1.0-.yaml}],
+            cmd_tags      => [q{Default: qc_eval_metric_-v1.1-.yaml}],
             documentation => q{File containing Qccollect evaluation metrics},
             is            => q{rw},
             isa           => Str,