+++ lib/MIP/Recipes/Install/Mip_scripts.pm
@@ -29,7 +29,7 @@ BEGIN {
     use base qw{ Exporter };
 
     # Set the version for version checking
-    our $VERSION = 1.17;
+    our $VERSION = 1.18;
 
     # Functions and variables which can be optionally exported
     our @EXPORT_OK = qw{ install_mip_scripts };
@@ -131,8 +131,8 @@ sub install_mip_scripts {
               mip_rd_dna_vcf_rerun_config.yaml
               mip_rd_rna_config.yaml
               program_test_cmds.yaml
-              qc_eval_metric_-v1.0-.yaml
-              qc_regexp_-v1.24-.yaml
+              qc_eval_metric_-v1.1-.yaml
+              qc_regexp_-v1.25-.yaml
               rank_model_-v1.29-.ini
               svrank_model_cmms_-v1.8-.ini
               }