+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Rename frequency_annotation to variant_annotation
 - Removed option to run sambamba markduplicates for markduplicates recipe
 - Added SpliceAI annotation
+- Collect and evaluate QC metrics generated in the RNA pipeline
 
 **Tools**
 - bcftools: 1.9=ha228f0b_4 -> 1.10.2-hd2cd319_0 (DNA)